level 2: nice logic, greedy good
This program was a great practice. Doing it by finding minimum lcm would be the most common approach but unfortunately it super slows the program. Exploring the number theory and then finding the right way to solve this one was super fun. Lastly the greedy condition(primes are odd except 2) to optimize loop was also awesome.