level 1: good practice
This [program was as easy as it could be. First time I read this question, I built the most complex and the worst algo possible - finding all permutations and comparing each to see if entered case is possible. At the end I was able to see that this was the stupidest way to do this program. Finally I revisited the program, observed some raw inputs and got a general condition for which this outputs YES. This was a good practice, as in future I will not try the same stupid approach of finding permutations in similar questions rated 800. Afterall who would think that a 800-rated coder knows the logic to find permutations.