Level 2: deep analysis
Although this question is rated 800, it is a really good one. Implementation is easy, but to implement it you need derive a conclusion. Deriving conclusion is the hard part in this program. I analyzed examples for this program till n=10 which was not really that easy as you also had to believe your intuitions in this one. The second hard part about this program was to store the output. I stored using strings but I believe using vectors might be a better option.